USB works MSC with SPI flash and DFU mode

+++ Src/drivers/SPI3.c
@@ -27,10 +27,9 @@ void SPI3_disable()
     SPI3->CR1 &= ~SPI_CR1_SPE; //	disabling SPI3
 }
 
-void SPI3_transmit(const uint8_t* data, uint8_t size)
+void SPI3_transmit(const uint8_t* data, int size)
 {
-
-    uint8_t  i = 0;
+    int  i = 0;
 
     while (i < size)
     {
